backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
requests
pillow
groq
acontext
pdf2image
faker
